requirements.txt
Django==4.2
djangorestframework
gunicorn
requests
dj-database-url
